fix: Blending backgrounds
